requirements-test.txt
-r requirements.txt

coverage==4.3.4
tox>=3.25.0
codecov>=2.0.0
django-multiselectfield@https://github.com/goinnn/django-multiselectfield/archive/07a5e442b211292a07c7abcf9402f3305c0fc292.zip
psycopg2-binary
pip-tools
